+++ lib/elixir_script/translator/kernel/defprotocol.ex
@@ -6,7 +6,6 @@ defmodule ElixirScript.Translator.Defprotocol do
   alias ElixirScript.Translator.Map
   alias ElixirScript.Translator.Utils
   alias ElixirScript.Translator.State
-  alias ElixirScript.ModuleSystems
   alias ElixirScript.Translator.Identifier
   require Logger
 
@@ -16,15 +15,14 @@ defmodule ElixirScript.Translator.Defprotocol do
   def make(name, functions, env) do
     { body, _ } = Defmodule.translate_body( {:__block__, [], [] }, env)
 
-    module_refs = ElixirScript.Translator.State.get_module_references(env.state, name)
-
-    {imports, body} = Defmodule.extract_imports_from_body(body)
-
     app_name = State.get_module(env.state, name).app
 
-    imports = imports ++ Defmodule.make_std_lib_import(env) ++
-      Defmodule.make_imports(app_name, module_refs, env) ++
-      [ElixirScript.ModuleSystems.import_module("Implementations", Utils.make_local_file_path(app_name, Utils.name_to_js_file_name(name) <> ".Defimpl", env)) ]
+    State.add_javascript_module_reference(
+      env.state,
+      env.module,
+      {:__aliases__, [], [:Implementations]},
+      Utils.make_local_file_path(app_name, Utils.name_to_js_file_name(name) <> ".Defimpl", env)
+    )
 
     object = process_spec_functions(functions)
     |> Enum.map(fn({key, value}) ->
@@ -79,11 +77,12 @@ defmodule ElixirScript.Translator.Defprotocol do
       )
     )
 
-    default = JS.export_default_declaration(JS.identifier(Utils.name_to_js_name(name)))
+    body = body ++ [declaration] ++ [implementations]
+    body = env.module_formatter.build(body, JS.identifier(Utils.name_to_js_name(name)), env)
 
     %{
       name: name,
-      body: imports ++ body ++ [declaration] ++ [implementations] ++ [default],
+      body: body,
       app_name: app_name
     }
   end
@@ -92,48 +91,5 @@ defmodule ElixirScript.Translator.Defprotocol do
     Enum.map(Keyword.keys(functions), fn(function_name) ->
       {function_name, JS.function_expression([], [], JS.block_statement([]))}
     end)
-  end
-
-  @doc """
-  Makes the protocol implementation module for the given implementation name.
-  This is used to consolidate all of the protocol implementations.
-  """
-  def make_defimpl(name, implementations \\ [], compiler_opts, env) do
-    imports = [ModuleSystems.import_module(:Elixir, Utils.make_local_file_path(:elixir, compiler_opts.core_path, compiler_opts.root, env))]
-
-    declarator = JS.variable_declarator(
-      JS.identifier("impls"),
-      JS.array_expression([])
-    )
-
-    declaration = JS.variable_declaration([declarator], :let)
-
-    default = JS.export_default_declaration(JS.identifier("impls"))
-
-    protocol_name = Atom.to_string(name)
-
-    app_name = State.get_module(env.state, name).app
-
-    body = Enum.flat_map(implementations, fn({impl_app_name, x}) ->
-      x = if is_atom(x), do: Atom.to_string(x), else: x
-      x = String.to_atom(protocol_name <> ".DefImpl." <> x)
-      name = Utils.name_to_js_name(x)
-      imports = ModuleSystems.import_module(name, Utils.make_local_file_path(impl_app_name, Utils.name_to_js_file_name(x), compiler_opts.root))
-      call = JS.call_expression(
-        JS.member_expression(
-          JS.identifier("impls"),
-          JS.identifier("push")
-        ),
-        [JS.identifier(name)]
-      )
-
-      [imports, call]
-    end)
-
-    %{
-      name: String.to_atom(protocol_name <> ".DefImpl"),
-      body: imports ++ [declaration] ++ body ++ [default],
-      app_name: app_name
-    }
   end
 end